@cyriltovena 2021-11-26
@jeschkies I rollback the buildModel change but kept the new way to compute required node.

I'll follow up with a better PR to improve performance later.